rename module to package
